+++ src/scripts/reload-modules.ts
@@ -1,6 +1,7 @@
 import * as vscode from 'vscode';
 
 import * as motionBuilderConsole from '../modules/motionbuilder-console';
+import * as logging from '../modules/logging';
 import * as utils from '../modules/utils';
 
 
@@ -10,6 +11,7 @@ interface IReloadModulesResponse {
     num_failed: number;
 }
 
+let isCommandRegistered = false;
 
 export async function main() {
     const disposableStatusMessage = vscode.window.setStatusBarMessage("$(sync~spin) Reloading modules...", 5000);
@@ -22,7 +24,7 @@ export async function main() {
         reloadScriptPath,
         "reload",
         {
-            "workspace_folders": workspaceFolders  // eslint-disable-line @typescript-eslint/naming-convention
+            "workspace_folders": workspaceFolders  // eslint-disable-line @typescrip6t-eslint/naming-convention
         }
     );
 
@@ -31,6 +33,33 @@ export async function main() {
     if (!response)
         return;
 
-    const statusBarItemIcon = response.num_failed > 0 ? "$(error)" : "$(check)";
-    vscode.window.setStatusBarMessage(`${statusBarItemIcon} Reloaded ${response.num_reloads} modules in ${response.time} ms`, 3500);
+    if (response.num_failed <= 0) {
+        vscode.window.setStatusBarMessage(`$(check) Reloaded ${response.num_reloads} modules in ${response.time} ms`, 3500);
+    }
+    else if (!isCommandRegistered) {
+        const statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 5);
+        statusBarItem.text = `$(error) Failed to reload ${response.num_failed} module${response.num_failed === 1 ? '' : 's'}`;
+        statusBarItem.command = "ue-python.showReloadErrorMessage";
+        statusBarItem.color = new vscode.ThemeColor('errorForeground');
+
+        const timeout = setTimeout(() => {
+            dispose();
+        }, 5000);
+
+        const commandDisposable = vscode.commands.registerCommand(statusBarItem.command, () => {
+            logging.getOutputChannel().show();
+            clearTimeout(timeout);
+            dispose();
+        });
+
+        const dispose = () => {
+            statusBarItem.dispose();
+            commandDisposable.dispose();
+            isCommandRegistered = false;
+        }
+
+        statusBarItem.show();
+
+        isCommandRegistered = true;
+    }
 }